Rename social files with legacy suffix
